Refactor external review artifact IO
